+++ src/services/campaignServices.ts
@@ -126,25 +126,37 @@ const campaignServices = {
     }
   },
 
-  createOrUpdateCampaign: async (formData: {
-    id?: string;
-    name: string;
-    location: string;
-    publicId: string;
-    imageUrl: string;
-    ownerId: number;
-    affiliation: number;
-    factions: { factionType: $Enums.FactionType; name: string }[];
-    players: Partial<User>[];
-  }) => {
+  createOrUpdateCampaign: async (
+    formData: {
+      id?: number;
+      name: string;
+      location: string;
+      picture: { imageUrl: string; publicId: string };
+      ownerId: number;
+      affiliation: number;
+      factions: { factionType: $Enums.FactionType; name: string }[];
+      players: Partial<User>[];
+    },
+    ownerId: number,
+  ) => {
+    if (formData.id) {
+      const campaign = await prisma.campaign.findUnique({
+        where: { id: formData.id },
+        select: { ownerId: true },
+      });
+      if (campaign?.ownerId !== ownerId) {
+        throw new Error('Only the owner of the campaign can update it');
+      }
+    }
+
     try {
       const campaign = await prisma.campaign.upsert({
-        where: { id: Number(formData?.id) || 0 },
+        where: { id: formData?.id || 0 },
         update: {
           name: formData.name,
           location: formData.location,
-          picture: { publicId: formData.publicId, imageUrl: formData.imageUrl },
-          ownerId: formData.ownerId,
+          picture: formData.picture,
+          ownerId,
           pendingPlayers: {
             connect: formData.players.map((user) => ({ id: user.id })),
           },
@@ -152,42 +164,39 @@ const campaignServices = {
         create: {
           name: formData.name,
           location: formData.location,
-          picture: { publicId: formData.publicId, imageUrl: formData.imageUrl },
-          ownerId: formData.ownerId,
+          picture: formData.picture,
+          ownerId,
           pendingPlayers: {
             connect: formData.players.map((user) => ({ id: user.id })),
           },
         },
       });
 
-      const createdFactions = await Promise.all(
-        formData.factions.map((faction) =>
-          prisma.faction.create({
+      console.log(formData);
+
+      if (!formData.id) {
+        const createdFactions = await Promise.all(
+          formData.factions.map((faction) =>
+            prisma.faction.create({
+              data: {
+                ...faction,
+                campaignId: campaign.id,
+              },
+            }),
+          ),
+        );
+
+        if (createdFactions.length >= 2) {
+          const [faction1, faction2] = createdFactions;
+
+          await prisma.affiliation.create({
             data: {
-              ...faction,
               campaignId: campaign.id,
-            },
-          }),
-        ),
-      );
-
-      if (createdFactions.length >= 2) {
-        const [faction1, faction2] = createdFactions;
-
-        await prisma.affiliation.createMany({
-          data: [
-            {
-              primaryFactionId: faction1.id,
-              secondaryFactionId: faction2.id,
+              factions: { connect: [{ id: faction1.id }, { id: faction2.id }] },
               value: formData.affiliation,
             },
-            {
-              primaryFactionId: faction2.id,
-              secondaryFactionId: faction1.id,
-              value: formData.affiliation,
-            },
-          ],
-        });
+          });
+        }
       }
 
       return campaign;
@@ -208,7 +217,7 @@ const campaignServices = {
 
       const inPending = campaign
         ? campaign.pendingPlayers.some(
-            (player: { id: number }) => player.id !== userId,
+            (player: { id: number }) => player.id === userId,
           )
         : [];
 
@@ -233,16 +242,29 @@ const campaignServices = {
     }
   },
 
-  deleteCampaign: async (campaignId: string) => {
+  deleteCampaign: async (campaignId: number, userId: number) => {
     try {
+      const campaign = await prisma.campaign.findUnique({
+        where: { id: campaignId },
+        select: { ownerId: true },
+      });
+
+      if (!campaign) {
+        throw new Error('Failed to find campaign');
+      }
+
+      if (campaign.ownerId !== userId) {
+        throw new Error('Only the owner of the campaign can delete it');
+      }
+
       await prisma.campaign.delete({
         where: {
           id: Number(campaignId),
         },
       });
-    } catch (error) {
+    } catch (error: any) {
       console.error(error);
-      throw new Error('Failed to delete campaign');
+      throw new Error(error.message || 'Failed to delete campaign');
     }
   },
 };
