+++ src/services/factionServices.ts
@@ -35,8 +35,7 @@ const factionServices = {
     id: number;
     name: string;
     background: { html: string; nodes: string };
-    publicId: string;
-    imageUrl: string;
+    picture: { imageUrl: string; publicId: string };
     affiliations: Affiliation[];
   }) => {
     try {
@@ -45,23 +44,9 @@ const factionServices = {
         data: {
           name: formData.name,
           background: formData.background,
-          picture: { publicId: formData.publicId, imageUrl: formData.imageUrl },
+          picture: formData.picture,
         },
       });
-
-      const affiliationUpdates = formData.affiliations.map((affiliation) => ({
-        id: affiliation.id,
-        value: affiliation.value,
-      }));
-
-      await Promise.all(
-        affiliationUpdates.map(({ id, value }) =>
-          prisma.affiliation.update({
-            where: { id },
-            data: { value },
-          }),
-        ),
-      );
 
       return faction;
     } catch (error) {
@@ -72,6 +57,25 @@ const factionServices = {
 
   deleteFaction: async (factionId: string) => {
     try {
+      const faction = await prisma.faction.findUnique({
+        where: {
+          id: Number(factionId),
+        },
+        select: {
+          affiliations: { select: { id: true } },
+        },
+      });
+
+      if (!faction) {
+        throw new Error('Failed to find faction');
+      }
+
+      await prisma.affiliation.deleteMany({
+        where: {
+          id: { in: faction.affiliations.map((affiliation) => affiliation.id) },
+        },
+      });
+
       await prisma.faction.delete({
         where: {
           id: Number(factionId),
